# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations/{organizationId}/catalog/items:
    get:
      operationId: listCatalogItems
      tags: [catalog]
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - { name: search, in: query, schema: { type: string } }
        - { name: type, in: query, schema: { $ref: "#/components/schemas/CatalogItemType" } }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/CatalogItemStatus" } }
        - { name: taxCategory, in: query, schema: { $ref: "#/components/schemas/TaxCategory" } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: sort, in: query, schema: { type: string, enum: [name, code, created, updated], default: name } }
      responses:
        "200": { description: Catalog items, content: { application/json: { schema: { $ref: "#/components/schemas/CatalogListResponse" } } } }
    post:
      operationId: createCatalogItem
      tags: [catalog]
      parameters: [{ $ref: "#/components/parameters/OrganizationId" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CatalogItemInput" } } } }
      responses:
        "201": { description: Created item, content: { application/json: { schema: { $ref: "#/components/schemas/CatalogItem" } } } }
  /organizations/{organizationId}/catalog/items/{itemId}:
    get:
      operationId: getCatalogItem
      tags: [catalog]
      parameters: [{ $ref: "#/components/parameters/OrganizationId" }, { $ref: "#/components/parameters/CatalogItemId" }]
      responses:
        "200": { description: Item, content: { application/json: { schema: { $ref: "#/components/schemas/CatalogItem" } } } }
    patch:
      operationId: updateCatalogItem
      tags: [catalog]
      parameters: [{ $ref: "#/components/parameters/OrganizationId" }, { $ref: "#/components/parameters/CatalogItemId" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CatalogItemUpdate" } } } }
      responses:
        "200": { description: Updated item, content: { application/json: { schema: { $ref: "#/components/schemas/CatalogItem" } } } }
  /organizations/{organizationId}/catalog/items/{itemId}/status:
    patch:
      operationId: updateCatalogItemStatus
      tags: [catalog]
      parameters: [{ $ref: "#/components/parameters/OrganizationId" }, { $ref: "#/components/parameters/CatalogItemId" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CatalogItemStatusInput" } } } }
      responses:
        "200": { description: Updated status, content: { application/json: { schema: { $ref: "#/components/schemas/CatalogItem" } } } }
  /organizations/{organizationId}/catalog/units:
    get:
      operationId: listCatalogUnits
      tags: [catalog]
      parameters: [{ $ref: "#/components/parameters/OrganizationId" }]
      responses:
        "200": { description: Units, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/CatalogUnit" } } } } }
  /organizations/{organizationId}/catalog/tax-definitions:
    get:
      operationId: listCatalogTaxDefinitions
      tags: [catalog]
      parameters: [{ $ref: "#/components/parameters/OrganizationId" }]
      responses:
        "200": { description: Tax definitions, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/TaxDefinition" } } } } }
  /organizations/{organizationId}/catalog/import/template:
    get:
      operationId: downloadCatalogImportTemplate
      tags: [catalog]
      parameters: [{ $ref: "#/components/parameters/OrganizationId" }]
      responses:
        "200": { description: CSV template, content: { text/csv: { schema: { type: string } } } }
  /organizations/{organizationId}/catalog/import/preview:
    post:
      operationId: previewCatalogImport
      tags: [catalog]
      parameters: [{ $ref: "#/components/parameters/OrganizationId" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CatalogImportInput" } } } }
      responses:
        "200": { description: Import validation preview, content: { application/json: { schema: { $ref: "#/components/schemas/CatalogImportPreview" } } } }
  /organizations/{organizationId}/catalog/import/confirm:
    post:
      operationId: confirmCatalogImport
      tags: [catalog]
      parameters: [{ $ref: "#/components/parameters/OrganizationId" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CatalogImportInput" } } } }
      responses:
        "201": { description: Import result, content: { application/json: { schema: { $ref: "#/components/schemas/CatalogImportResult" } } } }
  /organizations/{organizationId}/catalog/export:
    get:
      operationId: exportCatalogItems
      tags: [catalog]
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - { name: search, in: query, schema: { type: string } }
        - { name: type, in: query, schema: { $ref: "#/components/schemas/CatalogItemType" } }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/CatalogItemStatus" } }
        - { name: taxCategory, in: query, schema: { $ref: "#/components/schemas/TaxCategory" } }
      responses:
        "200": { description: CSV export, content: { text/csv: { schema: { type: string } } } }
components:
  parameters:
    CatalogItemId:
      name: itemId
      in: path
      required: true
      schema: { type: string, format: uuid }
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    CatalogItemType: { type: string, enum: [PRODUCT, SERVICE] }
    CatalogItemStatus: { type: string, enum: [ACTIVE, INACTIVE] }
    TaxCategory: { type: string, enum: [STANDARD, ZERO_RATED, EXEMPT, OUT_OF_SCOPE] }
    CatalogItemInput:
      type: object
      required: [type, name, nameAr, unitId, salesPrice, purchasePrice, taxCategory, taxRate]
      properties:
        type: { $ref: "#/components/schemas/CatalogItemType" }
        name: { type: string, minLength: 1 }
        nameAr: { type: string, minLength: 1 }
        description: { type: [string, "null"] }
        descriptionAr: { type: [string, "null"] }
        unitId: { type: string, format: uuid }
        salesPrice: { type: string, pattern: "^\\d+(\\.\\d{1,2})?$" }
        purchasePrice: { type: string, pattern: "^\\d+(\\.\\d{1,2})?$" }
        taxCategory: { $ref: "#/components/schemas/TaxCategory" }
        taxRate: { type: string, pattern: "^\\d+(\\.\\d{1,2})?$" }
        sku: { type: [string, "null"] }
        barcode: { type: [string, "null"] }
        trackInventory: { type: boolean, default: false }
    CatalogItemUpdate:
      allOf:
        - $ref: "#/components/schemas/CatalogItemInput"
    CatalogItemStatusInput:
      type: object
      required: [status]
      properties: { status: { $ref: "#/components/schemas/CatalogItemStatus" } }
    CatalogUnit:
      type: object
      required: [id, code, name, nameAr]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        nameAr: { type: string }
    CatalogItem:
      allOf:
        - $ref: "#/components/schemas/CatalogItemInput"
        - type: object
          required: [id, organizationId, code, status, createdAt, updatedAt]
          properties:
            id: { type: string, format: uuid }
            organizationId: { type: string, format: uuid }
            code: { type: string }
            status: { $ref: "#/components/schemas/CatalogItemStatus" }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    CatalogListResponse:
      type: object
      required: [items, page, pageSize, total, summary]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/CatalogItem" } }
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
        summary:
          type: object
          required: [total, products, services, active]
          properties:
            total: { type: integer }
            products: { type: integer }
            services: { type: integer }
            active: { type: integer }
    TaxDefinition:
      type: object
      required: [category, rate, name, nameAr]
      properties:
        category: { $ref: "#/components/schemas/TaxCategory" }
        rate: { type: string }
        name: { type: string }
        nameAr: { type: string }
    CatalogImportInput:
      type: object
      properties:
        csv: { type: [string, "null"] }
        rows: { type: array, items: { type: object, additionalProperties: true } }
    CatalogImportPreview:
      type: object
      required: [totalRows, validRows, errors, rows]
      properties:
        totalRows: { type: integer }
        validRows: { type: integer }
        errors: { type: array, items: { type: object, additionalProperties: true } }
        rows: { type: array, items: { type: object, additionalProperties: true } }
    CatalogImportResult:
      type: object
      required: [imported, skipped]
      properties:
        imported: { type: integer }
        skipped: { type: integer }
...