# lib/api-zod/src/generated/types/catalogImportInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CatalogImportInputRowsItem } from './catalogImportInputRowsItem';

export interface CatalogImportInput {
  /** @nullable */
  csv?: string | null;
  rows?: CatalogImportInputRowsItem[];
}