# lib/api-zod/src/generated/types/catalogImportPreview.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CatalogImportPreviewErrorsItem } from './catalogImportPreviewErrorsItem';
import type { CatalogImportPreviewRowsItem } from './catalogImportPreviewRowsItem';

export interface CatalogImportPreview {
  totalRows: number;
  validRows: number;
  errors: CatalogImportPreviewErrorsItem[];
  rows: CatalogImportPreviewRowsItem[];
}