# lib/api-zod/src/generated/types/catalogItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CatalogItemInput } from './catalogItemInput';
import type { CatalogItemStatus } from './catalogItemStatus';

export type CatalogItem = CatalogItemInput & {
  id: string;
  organizationId: string;
  code: string;
  status: CatalogItemStatus;
  createdAt: Date;
  updatedAt: Date;
};