# lib/api-zod/src/generated/types/catalogItemInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CatalogItemType } from './catalogItemType';
import type { TaxCategory } from './taxCategory';

export interface CatalogItemInput {
  type: CatalogItemType;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  nameAr: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  descriptionAr?: string | null;
  unitId: string;
  /** @pattern ^\d+(\.\d{1,2})?$ */
  salesPrice: string;
  /** @pattern ^\d+(\.\d{1,2})?$ */
  purchasePrice: string;
  taxCategory: TaxCategory;
  /** @pattern ^\d+(\.\d{1,2})?$ */
  taxRate: string;
  /** @nullable */
  sku?: string | null;
  /** @nullable */
  barcode?: string | null;
  trackInventory?: boolean;
}