# lib/api-zod/src/generated/types/catalogItemStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type CatalogItemStatus = typeof CatalogItemStatus[keyof typeof CatalogItemStatus];


export const CatalogItemStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
} as const;