# lib/api-zod/src/generated/types/catalogItemType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type CatalogItemType = typeof CatalogItemType[keyof typeof CatalogItemType];


export const CatalogItemType = {
  PRODUCT: 'PRODUCT',
  SERVICE: 'SERVICE',
} as const;