# lib/api-zod/src/generated/types/catalogItemUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CatalogItemInput } from './catalogItemInput';

export type CatalogItemUpdate = CatalogItemInput;