# lib/api-zod/src/generated/types/catalogListResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CatalogItem } from './catalogItem';
import type { CatalogListResponseSummary } from './catalogListResponseSummary';

export interface CatalogListResponse {
  items: CatalogItem[];
  page: number;
  pageSize: number;
  total: number;
  summary: CatalogListResponseSummary;
}