# lib/api-zod/src/generated/types/catalogUnit.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CatalogUnit {
  id: string;
  code: string;
  name: string;
  nameAr: string;
}