# lib/api-zod/src/generated/types/taxDefinition.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaxCategory } from './taxCategory';

export interface TaxDefinition {
  category: TaxCategory;
  rate: string;
  name: string;
  nameAr: string;
}